Add a Dockerfile and some related config
This isn't fully done right now. I think this is enough to work in prod
(where we have the secrets we need -- we'll find out), but definitely
not in dev yet.

There are also some other things that would eventually need to do before
we can replace coma with this, at least:

* The background worker doesn't run in this new setup yet
* Static files are still served from static.ocf.b.e which is coma