Parallelize tests run and bump requirements
This cuts down teest time from around 120 seconds using one core on
reaper to around 25 seconds using all cores at once. My only real
concern with this is that if a lot of builds of ocfweb run at once that
it'll overwhelm a machine, but I think it'll be ok, especially since all
tests finish in under 30 seconds anyway.