Bump pre-commit hooks
This relies on #443, since one of the hooks removes trailing whitespace
from markdown as well. This is fine, since I believe this is the correct
path to take, but trailing whitespace in markdown does have a functional
use in inserting hard line breaks, so the feature in #443 is needed to
replace that functionality.