add a starter task for making a PR
Making pull requests is a critical part of contributing the OCF. This
starter task will walk people though the process in a lot of detail so
that they can start to get comfortable. It also makes an observable
change to the actual website, so that the task feels less like a
meaningless exercise.