Upgrade pygments, don't try to guess a lexer
With newer versions of pygments, trying to guess a lexer was giving some
lexers like MIMELexer, TeraTermLexer, or TransactSqlLexer that are
definitely not what we want. Instead of trying to guess, just default to
text unless a specific language is passed in. If a specific language is
used and there is no lexer that matches, don't catch the error so it
bubbles up and is clear to the person writing the page that they have
chosen an invalid lexer option.